Fix warnings of inconsistent types in AudioRenderer
Declare |max_cached_frames| and |max_frames_per_append| as int instead
of size_t as most of the places they are used are signed operations.

+++ starboard/android/shared/player_components_impl.cc
@@ -85,8 +85,8 @@ class PlayerComponentsImpl : public PlayerComponents {
     video_render_algorithm->reset(new VideoRenderAlgorithmImpl);
   }
 
-  void GetAudioRendererParams(size_t* max_cached_frames,
-                              size_t* max_frames_per_append) const override {
+  void GetAudioRendererParams(int* max_cached_frames,
+                              int* max_frames_per_append) const override {
     SB_DCHECK(max_cached_frames);
     SB_DCHECK(max_frames_per_append);
